Moved handling of character deletion to the character service.
