+++ src/login/providers/account_provider_interface.h
@@ -25,7 +25,7 @@
 #include <string>
 #include <tuple>
 
-#include "login/account.h"
+#include "swganh/login/account.h"
 
 namespace login {
 namespace providers {
@@ -34,7 +34,7 @@ class AccountProviderInterface {
 public:
     virtual ~AccountProviderInterface() {}
 
-    virtual std::shared_ptr<login::Account> FindByUsername(std::string username) = 0;
+    virtual std::shared_ptr<swganh::login::Account> FindByUsername(std::string username) = 0;
     virtual uint32_t FindBySessionKey(const std::string& session_key) = 0;
     virtual bool CreateAccountSession(uint32_t account_id, const std::string& session_key) = 0;
 };
