+++ src/login/providers/mysql_account_provider.h
@@ -34,7 +34,7 @@ public:
     explicit MysqlAccountProvider(std::shared_ptr<anh::database::DatabaseManagerInterface> db_manager);
     ~MysqlAccountProvider();
 
-    std::shared_ptr<login::Account> FindByUsername(std::string username);
+    std::shared_ptr<swganh::login::Account> FindByUsername(std::string username);
     uint32_t FindBySessionKey(const std::string& session_key);
     bool CreateAccountSession(uint32_t account_id, const std::string& session_key);
 private:
